+++ src/gui/fileService/FileService.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019 Damir Porobic <damir.porobic@gmx.com>
+ * Copyright (C) 2020 Damir Porobic <damir.porobic@gmx.com>
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -17,34 +17,10 @@
  * Boston, MA 02110-1301, USA.
  */
 
-#ifndef KSNIP_UPLOADOPERATION_H
-#define KSNIP_UPLOADOPERATION_H
+#include "FileService.h"
 
-#include <QCoreApplication>
-#include <QImage>
-
-#include <utility>
-
-#include "src/backend/uploader/IUploader.h"
-#include "src/backend/config/KsnipConfigProvider.h"
-#include "src/gui/messageBoxService/MessageBoxService.h"
-
-class UploadOperation : public QObject
+bool FileService::remove(const QString &path)
 {
-	Q_OBJECT
-public:
-	UploadOperation(QImage image, IUploader *uploader);
-	~UploadOperation() override;
-	bool execute();
-
-private:
-	KsnipConfig *mConfig;
-	IUploader *mUploader;
-	QImage mImage;
-	IMessageBoxService *mMessageBoxService;
-
-	bool proceedWithUpload() const;
-	bool askIfCanProceedWithUpload() const;
-};
-
-#endif //KSNIP_UPLOADOPERATION_H
+	QFile file(path);
+	return file.remove();
+}